22251226黄建龙/Project02/requirements.txt
numpy
pyopengl==3.1.6
PyOpenGL_accelerate==3.1.6
pillow==9.2.0
